+++ Tests/Tests.Shared/TestObjects.cs
@@ -205,6 +205,18 @@ namespace Tests
     {
         [PrimaryKey]
         public string StringProperty { get; set; }
+
+        public string Value { get; set; }
+    }
+
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass")]
+    public class RequiredPrimaryKeyStringObject : RealmObject
+    {
+        [PrimaryKey]
+        [Required]
+        public string StringProperty { get; set; }
+
+        public string Value { get; set; }
     }
 
     [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleClass")]